Remove bugged scenarios from list
Bugged scenarios are removed from the list. Sitting scenarios have been added to a new list which is checked when playing a scenario, when the scenario is in that list, it will play the scenario at different coordinates relative to the player, rather than the current position of the player.