Loading & Saving Vehicles!
Saving and loading saved vehicles now works(-ish). Currently only the model display name, model hash and a custom save name are saved, can be loaded and can be deleted. The vehicle mods, colors, liveries and other customization features are not yet saved. But that's very easy to add because the base for saving and loading (dynamically) is ready, albeit a bit hacky.