Add more death messages for the notifications.
Fix vehicle (weapons) not triggering notifications because the entity didn't match a player. Extra checks should now make sure this still registers as a proper death and should send notifications, even if the entity could not be found or is invalid in some other way. (basically all 'else' cases are now properly handeled).